Компилируется Device без предупреждений

+++ sources/Device/src/Hardware/Modules/BME280/BME280.cpp
@@ -53,7 +53,7 @@ bool BME280::GetMeasures(float* temp, float* pressure, float* humidity)
         return false;
     }
 
-    timeNext += TIME_MEASURE + (std::rand() % 100);
+    timeNext += (uint)TIME_MEASURE + (uint)(std::rand() % 100);
 
 #ifdef IN_MODE_TEST
